测试提交订单，（提交中文会变成乱码，未解决,单价表应该设置一个排名字段）

+++ src/main/java/com/yipeng/bill/bms/dao/BillMapper.java
@@ -8,7 +8,7 @@ import java.util.Map;
 public interface BillMapper {
     int deleteByPrimaryKey(Long id);
 
-    int insert(Bill record);
+    Long insert(Bill record);
 
     int insertSelective(Bill record);
